issue warning on undeclared function type

+++ tests/functional/dup_func_decl.bas
@@ -1,6 +1,6 @@
 
 
-declare function f()
+declare function f() as float
 
 declare function f()
